refactor: 공통 훅 위치 변경
