+++ src/hooks/date-picker/components/SelectDate.tsx
@@ -1,11 +1,13 @@
-import CalenderBox from "@containers/home/HomeContainer/view/Calender/boxes/CalenderBox";
+import CalenderBox from "@containers/home/HomeContainer/view/Calender/boxes/CalenderBox.tsx";
 import { DateCalendar } from "@mui/x-date-pickers";
 import moment from "moment";
 import { Dispatch, SetStateAction } from "react";
+
 interface SelectDateProps {
   value: string;
   setValue: Dispatch<SetStateAction<string>>;
 }
+
 function SelectDate({ value, setValue }: SelectDateProps) {
   return (
     <CalenderBox dateHeight={50} dateSize={32} week={6}>
