+++ src/hooks/date-picker/hooks/useTimePicker.tsx
@@ -1,12 +1,17 @@
 import { useOverlay } from "@hooks/use-overlay/useOverlay.tsx";
-import DatePicker from "@components/layouts/date-picker/components/DatePicker.tsx";
+import TimePicker from "@hooks/date-picker/components/TimePicker.tsx";
 
-export const useDatePicker = () => {
+export const useTimePicker = () => {
   const { openOverlay, closeOverlay } = useOverlay();
-  const openDatePicker = (): Promise<string> => {
+  const openTimePicker = ({
+    defaultTime,
+  }: {
+    defaultTime: string;
+  }): Promise<string> => {
     return new Promise((resolve) => {
       openOverlay(
-        <DatePicker
+        <TimePicker
+          defaultTime={defaultTime}
           onClickApprove={(answer) => {
             resolve(answer);
             closeOverlay();
@@ -20,5 +25,5 @@ export const useDatePicker = () => {
     });
   };
 
-  return { openDatePicker, closeDatePicker: closeOverlay };
+  return { openTimePicker, closeTimePicker: closeOverlay };
 };
